werer

+++ api/controllers/base/actions/UpdateAction.php
@@ -48,6 +48,7 @@ class UpdateAction extends Action
         if (!$model->validate()) {
             $errors[] = $model->errors;
         }
+        $rows = [];
         
         if (isset($model->details)) {
             foreach($model->details as $key => $detail) {
@@ -100,13 +101,15 @@ class UpdateAction extends Action
                     if ($model->update(false)) {
                         if (isset($model->details)) {
                             foreach($model->details as $key => $detail) {
-                                $detail::deleteAll([$detail::$ownerForeignKey => $model->id]);
-                                foreach($rows[$key] as $k => $val) {
-                                    $rows[$key][$k][$detail::$ownerForeignKey] = $model->id;
-                                }
-                        /*        \Yii::info(array_keys($rows[$key][0]));
-                                \Yii::info(array_values($rows[$key]));  */
-                                \Yii::$app->db->createCommand()->batchInsert($detail::tableName(), array_keys($rows[$key][0]), array_values($rows[$key]))->execute();
+                                if (isset($rows[$key])) {
+                                    $detail::deleteAll([$detail::$ownerForeignKey => $model->id]);
+                                    foreach($rows[$key] as $k => $val) {
+                                        $rows[$key][$k][$detail::$ownerForeignKey] = $model->id;
+                                    }
+                            /*        \Yii::info(array_keys($rows[$key][0]));
+                                    \Yii::info(array_values($rows[$key]));  */
+                                    \Yii::$app->db->createCommand()->batchInsert($detail::tableName(), array_keys($rows[$key][0]), array_values($rows[$key]))->execute();
+                                }    
                             }
                         }
